requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
weasyprint
pytest
Flask-Testing
python-dotenv
